+++ changelog.md
@@ -1,6 +1,7 @@
 19.1.0
   - added connections field to MetatraderAccount model to track CopyFactory / risk management API connection status in addition to MetaApi connection status
   - added RPC method to retrieve current server time and a terminal state method to retrieve latest quote time
+  - added riskManagementApiEnabled field to MetatraderAccount model
 
 19.0.1
   - fixed recording historyOrders to disk